+++ Packages/com.nickmaltbie.openkcc/OpenKCC/Utils/IColliderCast.cs
@@ -74,22 +74,13 @@ namespace nickmaltbie.OpenKCC.Utils
         Vector3 PushOutOverlapping(Vector3 position, Quaternion rotation, float maxDistance);
 
         /// <summary>
-        /// Draw a given mesh with the gizmos in the world.
-        /// </summary>
-        /// <param name="outlineColor">Color to draw outline of the mesh.</param>
-        /// <param name="fillColor">Color to draw the fill of the mesh.</param>
-        /// <param name="position">Position to draw the mesh in.</param>
-        /// <param name="rotation">Rotation to draw the mesh in.</param>
-        void DrawMeshGizmo(Color outlineColor, Color fillColor, Vector3 position, Quaternion rotation);
-
-        /// <summary>
-        /// Check if there is a vertical step ahead of the character at a given location.
+        /// Do a raycast in a given direction ignoring this object.
         /// </summary>
         /// <param name="source">Source point to check from.</param>
         /// <param name="direction">Direction to search for step.</param>
         /// <param name="distance">Distance to search for step ahead of player.</param>
         /// <param name="stepHit">Information bout step hit ahead.</param>
-        /// <returns>Step hit information ahead of player.</returns>
-        bool CheckVerticalStepAhead(Vector3 source, Vector3 direction, float distance, out IRaycastHit stepHit);
+        /// <returns>Is something ahead hit.</returns>
+        bool DoRaycastInDirection(Vector3 source, Vector3 direction, float distance, out IRaycastHit stepHit);
     }
 }